+++ cmd/service.go
@@ -14,5 +14,6 @@ func NewCmdService(w io.Writer, rf *RootFlags) *cobra.Command {
 	}
 
 	cmd.AddCommand(NewCmdRunService(w, rf))
+	cmd.AddCommand(NewCmdListEndpoints(w, rf))
 	return cmd
 }